requirements.txt
dotenv==0.9.9
fpdf2==2.8.2
Faker==37.4.0
openpyxl==3.1.5
pandas==2.3.0
python-docx==1.1.2
supabase==2.16.0